native: move handlers + props from GroupChannelsScreen out to useGroupContext
Fixes TLON-2451
